requirements.txt
httpx>=0.27.0
colorama>=0.4.6
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.0
